Support non existent topics in KCBQ.
